coprocessor:share the Error struct in mod codec and expr
